backtrace: fix warnings on 64 bit
src/backtrace.c: In function 'translate_addresses_buf':
src/backtrace.c:158: warning: field width should have type 'int', but argument 4 has type 'long unsigned int'
src/backtrace.c:158: warning: field width should have type 'int', but argument 4 has type 'long unsigned int'

+++ src/backtrace.c
@@ -134,7 +134,7 @@ static char **translate_addresses_buf(const char *obj_file, const void *obj_base
 
 			if (!found) {
 				total += snprintf(buf, len, "0x%0*lx in \?\?() from %s at \?\?:0",
-						  2 * sizeof(void *),
+						  (int)(2 * sizeof(void *)),
 						  (long unsigned int)(obj_base + addr[naddr - 1]),
 						  obj_file) + 1;
 			} else {
@@ -151,7 +151,7 @@ static char **translate_addresses_buf(const char *obj_file, const void *obj_base
 						filename = h + 1;
 				}
 				total += snprintf(buf, len, "0x%0*lx in %s() from %s at %s:%u",
-						  2 * sizeof(void *),
+						  (int)(2 * sizeof(void *)),
 						  (long unsigned int)(obj_base + addr[naddr - 1]),
 						  name, obj_file,
 						  filename ? filename : "??",
